[Refactor] Move analyze logic from ColumnDef AST to ColumnDefAnalyzer
