+++ fe/fe-core/src/main/java/com/starrocks/connector/iceberg/IcebergAlterTableExecutor.java
@@ -144,7 +144,7 @@ public class IcebergAlterTableExecutor extends ConnectorAlterTableExecutor {
         actions.add(() -> {
             UpdateSchema updateSchema = this.transaction.updateSchema();
             ColumnPosition pos = clause.getColPos();
-            Column column = clause.getColumnDef().toColumn(null);
+            Column column = Column.fromColumnDef(null, clause.getColumnDef());
 
             // All non-partition columns must use NULL as the default value.
             if (!column.isAllowNull()) {
@@ -178,7 +178,7 @@ public class IcebergAlterTableExecutor extends ConnectorAlterTableExecutor {
             List<Column> columns = clause
                     .getColumnDefs()
                     .stream()
-                    .map(columnDef -> columnDef.toColumn(null))
+                    .map(columnDef -> Column.fromColumnDef(null, columnDef))
                     .collect(Collectors.toList());
 
             for (Column column : columns) {
@@ -220,7 +220,7 @@ public class IcebergAlterTableExecutor extends ConnectorAlterTableExecutor {
         actions.add(() -> {
             UpdateSchema updateSchema = this.transaction.updateSchema();
             ColumnPosition colPos = clause.getColPos();
-            Column column = clause.getColumnDef().toColumn(null);
+            Column column = Column.fromColumnDef(null, clause.getColumnDef());
             org.apache.iceberg.types.Type colType = toIcebergColumnType(column.getType());
 
             // UPDATE column type
